Deleted `VersionAndBranchFinder`, moved its caching logic into `ExecuteCore` and replaced all usage of the former with the latter. Also made `ExecuteCore` non-static.
