+++ src/GitVersionCore.Tests/GitVersionCore.Tests.csproj
@@ -178,6 +178,7 @@
     <Compile Include="TestFileSystem.cs" />
     <Compile Include="TestStream.cs" />
     <Compile Include="VariableProviderTests.cs" />
+    <Compile Include="ExecuteCoreTests.cs" />
     <Compile Include="VersionCalculation\BaseVersionCalculatorTests.cs" />
     <Compile Include="VersionCalculation\NextVersionCalculatorTests.cs" />
     <Compile Include="VersionCalculation\Strategies\ConfigNextVersionBaseVersionStrategyTests.cs" />
