+++ src/GitVersionExe/SpecifiedArgumentRunner.cs
@@ -1,13 +1,23 @@
 namespace GitVersion
 {
     using System;
+    using System.Collections.Concurrent;
     using System.Collections.Generic;
     using System.Linq;
+
     using GitVersion.Helpers;
 
     class SpecifiedArgumentRunner
     {
         const string MsBuild = @"c:\Windows\Microsoft.NET\Framework\v4.0.30319\msbuild.exe";
+        static readonly ConcurrentDictionary<string, VersionVariables> versionCache;
+
+
+        static SpecifiedArgumentRunner()
+        {
+            versionCache = new ConcurrentDictionary<string, VersionVariables>();
+        }
+
 
         public static void Run(Arguments arguments, IFileSystem fileSystem)
         {
@@ -19,7 +29,8 @@ namespace GitVersion
             var targetBranch = arguments.TargetBranch;
             var commitId = arguments.CommitId;
 
-            var variables = ExecuteCore.ExecuteGitVersion(fileSystem, targetUrl, dynamicRepositoryLocation, authentication, targetBranch, noFetch, targetPath, commitId);
+            var executeCore = new ExecuteCore(fileSystem, versionCache.GetOrAdd);
+            var variables = executeCore.ExecuteGitVersion(targetUrl, dynamicRepositoryLocation, authentication, targetBranch, noFetch, targetPath, commitId);
 
             if (arguments.Output == OutputType.BuildServer)
             {
@@ -66,9 +77,13 @@ namespace GitVersion
             }
         }
 
+
         static bool RunMsBuildIfNeeded(Arguments args, string workingDirectory, VersionVariables variables)
         {
-            if (string.IsNullOrEmpty(args.Proj)) return false;
+            if (string.IsNullOrEmpty(args.Proj))
+            {
+                return false;
+            }
 
             Logger.WriteInfo(string.Format("Launching {0} \"{1}\" {2}", MsBuild, args.Proj, args.ProjArgs));
             var results = ProcessHelper.Run(
@@ -77,14 +92,20 @@ namespace GitVersion
                 GetEnvironmentalVariables(variables));
 
             if (results != 0)
+            {
                 throw new WarningException("MsBuild execution failed, non-zero return code");
+            }
 
             return true;
         }
 
+
         static bool RunExecCommandIfNeeded(Arguments args, string workingDirectory, VersionVariables variables)
         {
-            if (string.IsNullOrEmpty(args.Exec)) return false;
+            if (string.IsNullOrEmpty(args.Exec))
+            {
+                return false;
+            }
 
             Logger.WriteInfo(string.Format("Launching {0} {1}", args.Exec, args.ExecArgs));
             var results = ProcessHelper.Run(
@@ -92,10 +113,13 @@ namespace GitVersion
                 null, args.Exec, args.ExecArgs, workingDirectory,
                 GetEnvironmentalVariables(variables));
             if (results != 0)
+            {
                 throw new WarningException(string.Format("Execution of {0} failed, non-zero return code", args.Exec));
+            }
 
             return true;
         }
+
 
         static KeyValuePair<string, string>[] GetEnvironmentalVariables(VersionVariables variables)
         {
