+++ src/GitVersionTask.Tests/GitVersionTask.Tests.csproj
@@ -131,7 +131,6 @@
     <Compile Include="Helpers\SelfCleaningDirectory.cs" />
     <Compile Include="Mocks\MockTaskItem.cs" />
     <Compile Include="ModuleInitializer.cs" />
-    <Compile Include="VersionAndBranchFinderTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
